Extract SyncStats into src/common
