+++ src/engagement_db_coda_sync/sync_stats.py
@@ -1,36 +1,30 @@
 from core_data_modules.logging import Logger
 
+from src.common.sync_stats import SyncStats
+
 log = Logger(__name__)
 
 
 class CodaSyncEvents:
     READ_MESSAGE_FROM_ENGAGEMENT_DB = "read_message_from_engagement_db"
     SET_CODA_ID = "set_coda_id"
+    READ_MESSAGE_FROM_CODA = "read_message_from_coda"
     ADD_MESSAGE_TO_CODA = "add_message_to_coda"
     LABELS_MATCH = "labels_match"
     UPDATE_ENGAGEMENT_DB_LABELS = "update_engagement_db_labels"
     WS_CORRECTION = "ws_correction"
 
 
-class CodaSyncStats:
+class CodaSyncStats(SyncStats):
     def __init__(self):
-        self.event_counts = {
+        super().__init__({
             CodaSyncEvents.READ_MESSAGE_FROM_ENGAGEMENT_DB: 0,
             CodaSyncEvents.SET_CODA_ID: 0,
             CodaSyncEvents.ADD_MESSAGE_TO_CODA: 0,
             CodaSyncEvents.LABELS_MATCH: 0,
             CodaSyncEvents.UPDATE_ENGAGEMENT_DB_LABELS: 0,
             CodaSyncEvents.WS_CORRECTION: 0
-        }
-
-    def add_event(self, event):
-        if event not in self.event_counts:
-            self.event_counts[event] = 0
-        self.event_counts[event] += 1
-
-    def add_stats(self, stats):
-        for k, v in stats.event_counts.items():
-            self.event_counts[k] += v
+        })
 
     def print_summary(self):
         log.info(f"Messages read from engagement db: {self.event_counts[CodaSyncEvents.READ_MESSAGE_FROM_ENGAGEMENT_DB]}")
